Dialects plugin: Consider slop in deduplication logic, allow configuring dialect/partition caps
