+++ all/plugin-dialects/dialectsplugin/__init__.py
@@ -1,0 +1,1 @@
+"""Package for internals of the dialects plugin."""
