+++ all/plugin-dialects/dialectsplugin/similarity.py
@@ -1,7 +1,4 @@
-import abc
-import dataclasses
-import sys
-from typing import Collection, List, Mapping, Optional, Sequence, Union, Tuple
+from typing import Sequence, Tuple
 
 import numpy as np
 import numpy.typing as npt
