--- all/plugin-dialects/partitioning/__init__.py
@@ -1,17 +1,0 @@
-"""Package for internals of the dialects plugin."""
-from partitioning.partitions import best_partitions, target_files_from_cnf
-from partitioning.similarity import similar_features_to, file_distribution_similarity, bernoulli_entropy_weights, features_attributable_to, file_distribution_attributable_risk
-from partitioning.quality import quality_metrics
-from partitioning.filtering import TargetRestrictionMode
-
-__all__ = [
-    best_partitions,
-    target_files_from_cnf,
-    similar_features_to,
-    file_distribution_similarity,
-    features_attributable_to,
-    file_distribution_attributable_risk,
-    bernoulli_entropy_weights,
-    quality_metrics,
-    TargetRestrictionMode,
-]
