+++ all/plugin-dialects/requirements.txt
@@ -2,4 +2,3 @@ numpy
 scikit-learn
 scipy
 typer
-
